+++ test/unit/test_states.py
@@ -1,7 +1,10 @@
 """Unit test suite for ``rhodes.states``."""
 import pytest
 
-from rhodes.states import Pass
+import jsonpath_rw
+
+from rhodes.states import Pass, StateMachine
+from rhodes.structures import JsonPath
 
 pytestmark = [pytest.mark.local, pytest.mark.functional]
 
@@ -18,3 +21,31 @@ def test_invalid_end():
         Pass("Foo", End=False)
 
     excinfo.match("If 'End' is set, value must be True")
+
+
+@pytest.mark.parametrize(
+    "cls, name, value",
+    ((Pass, "InputPath", JsonPath("$")), (Pass, "OutputPath", JsonPath("$")), (Pass, "ResultPath", JsonPath("$"))),
+)
+def test_default_value(cls, name, value):
+    instance = cls("Foo")
+
+    test = getattr(instance, name)
+
+    assert test == value
+
+
+@pytest.mark.parametrize("source_result, requested_promotion, expected_input", (
+        ("$.foo.bar", "@.baz", "$.foo.bar.baz"),
+        ("$", "@.bar.baz", "$.bar.baz"),
+))
+@pytest.mark.parametrize("path_reader", (lambda x: x, jsonpath_rw.parse, JsonPath))
+def test_promote(path_reader, source_result, requested_promotion, expected_input):
+    machine = StateMachine()
+    starter = machine.start_with(Pass("Foo", ResultPath=JsonPath(source_result)))
+    test = starter.promote(path_reader(requested_promotion))
+
+    assert test.title == "Foo-PromoteResult"
+    assert test.InputPath == JsonPath(expected_input)
+    assert test.ResultPath == JsonPath(source_result)
+    assert test.member_of is machine